# Issue 4274: Write padding more efficiently 