@RaulGracia 2019-10-12
I have executed some preliminary experiments on EKS and I don't see much of a difference in terms of performance (specially given the variability):
```
/bin/pravega-benchmark -controller tcp://10.100.110.33:9090 -enableConnectionPooling False -segments 1 -producers 1 -stream stream1segment -size 100 -time 60
(this PR)
INFO io.pravega.perf.PerfStats - 1109493 records Writing, 18743.019 records/sec, 100 bytes record size, 1.79 MiB/sec, 6.9 ms avg latency, 61.0 ms max latency, 7 ms 50th, 8 ms 75th, 9 ms 95th, 11 ms 99th, 16 ms 99.9th, 24 ms 99.99th.
INFO io.pravega.perf.PerfStats - 1239390 records Writing, 20956.173 records/sec, 100 bytes record size, 2.00 MiB/sec, 7.0 ms avg latency, 73.0 ms max latency, 7 ms 50th, 8 ms 75th, 9 ms 95th, 12 ms 99th, 17 ms 99.9th, 18 ms 99.99th
INFO io.pravega.perf.PerfStats - 1123416 records Writing, 18981.110 records/sec, 100 bytes record size, 1.81 MiB/sec, 7.0 ms avg latency, 86.0 ms max latency, 7 ms 50th, 8 ms 75th, 9 ms 95th, 11 ms 99th, 51 ms 99.9th, 76 ms 99.99th.

(master)
INFO io.pravega.perf.PerfStats - 1098745 records Writing, 18556.747 records/sec, 100 bytes record size, 1.77 MiB/sec, 7.0 ms avg latency, 60.0 ms max latency, 7 ms 50th, 8 ms 75th, 9 ms 95th, 10 ms 99th, 13 ms 99.9th, 17 ms 99.99th.
INFO io.pravega.perf.PerfStats - 1132750 records Writing, 19132.675 records/sec, 100 bytes record size, 1.82 MiB/sec, 6.9 ms avg latency, 78.0 ms max latency, 7 ms 50th, 8 ms 75th, 9 ms 95th, 12 ms 99th, 37 ms 99.9th, 71 ms 99.99th
INFO io.pravega.perf.PerfStats - 1128576 records Writing, 19060.242 records/sec, 100 bytes record size, 1.82 MiB/sec, 7.1 ms avg latency, 146.0 ms max latency, 7 ms 50th, 8 ms 75th, 9 ms 95th, 11 ms 99th, 87 ms 99.9th, 136 ms 99.99th.
INFO io.pravega.perf.PerfStats - 6890219 records Writing, 116603.527 records/sec, 100 bytes record size, 11.12 MiB/sec, 16.4 ms avg latency, 958.0 ms max latency, 9 ms 50th, 11 ms 75th, 39 ms 95th, 148 ms 99th, 812 ms 99.9th, 915 ms 99.99th.
```

```
./bin/pravega-benchmark -controller tcp://10.100.110.33:9090 -enableConnectionPooling False -segments 4 -producers 10 -stream stream4segment -size 100 -time 60
(this PR)
INFO io.pravega.perf.PerfStats - 6902575 records Writing, 117578.697 records/sec, 100 bytes record size, 11.21 MiB/sec, 9.6 ms avg latency, 97.0 ms max latency, 8 ms 50th, 11 ms 75th, 19 ms 95th, 35 ms 99th, 57 ms 99.9th, 83 ms 99.99th.
INFO io.pravega.perf.PerfStats - 6973056 records Writing, 118033.346 records/sec, 100 bytes record size, 11.26 MiB/sec, 9.8 ms avg latency, 135.0 ms max latency, 8 ms 50th, 11 ms 75th, 19 ms 95th, 40 ms 99th, 80 ms 99.9th, 110 ms 99.99th.
INFO io.pravega.perf.PerfStats - 6872002 records Writing, 116366.133 records/sec, 100 bytes record size, 11.10 MiB/sec, 10.4 ms avg latency, 98.0 ms max latency, 8 ms 50th, 11 ms 75th, 24 ms 95th, 57 ms 99th, 74 ms 99.9th, 88 ms 99.99th.

(master)
INFO io.pravega.perf.PerfStats - 6958852 records Writing, 117802.885 records/sec, 100 bytes record size, 11.23 MiB/sec, 9.5 ms avg latency, 99.0 ms max latency, 8 ms 50th, 11 ms 75th, 17 ms 95th, 34 ms 99th, 63 ms 99.9th, 82 ms 99.99th.
INFO io.pravega.perf.PerfStats - 6952792 records Writing, 117734.180 records/sec, 100 bytes record size, 11.23 MiB/sec, 10.2 ms avg latency, 120.0 ms max latency, 8 ms 50th, 11 ms 75th, 22 ms 95th, 49 ms 99th, 84 ms 99.9th, 103 ms 99.99th.
```

But if this PR also fixes a correctness problem in the tests, then that's fine.