@tkaitchuck 2019-10-14
The code change does not affect correctness. If there is no perf advantage, I will roll back that part of the change and make a PR on just the test changes.